fix(webwolf): contain uploaded filenames within user directory

+++ src/main/java/org/owasp/webgoat/webwolf/FileServer.java
@@ -69,14 +69,23 @@ public class FileServer {
     var username = authentication.getName();
     var destinationDir = new File(fileLocation, username);
     destinationDir.mkdirs();
+    var originalFilename = multipartFile.getOriginalFilename();
+    if (originalFilename == null) {
+      throw new IllegalArgumentException("Invalid filename");
+    }
+    var canonicalDestinationDir = destinationDir.getCanonicalFile().toPath();
+    var destinationFile = new File(destinationDir, originalFilename).getCanonicalFile().toPath();
+    if (!destinationFile.startsWith(canonicalDestinationDir)
+        || destinationFile.equals(canonicalDestinationDir)) {
+      throw new IllegalArgumentException("Invalid filename");
+    }
     // DO NOT use multipartFile.transferTo(), see
     // https://stackoverflow.com/questions/60336929/java-nio-file-nosuchfileexception-when-file-transferto-is-called
     try (InputStream is = multipartFile.getInputStream()) {
-      var destinationFile = destinationDir.toPath().resolve(multipartFile.getOriginalFilename());
       Files.deleteIfExists(destinationFile);
       Files.copy(is, destinationFile);
     }
-    log.debug("File saved to {}", new File(destinationDir, multipartFile.getOriginalFilename()));
+    log.debug("File saved to {}", destinationFile);
 
     return new ModelAndView(
         new RedirectView("files", true),
